Refactor API structure: consolidate v1 and v2 routers, update schema imports, and remove obsolete files
