+++ src/api/v1/router.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Depends, HTTPException
 from fastapi.responses import StreamingResponse
 
 from src.api.v1.ollama_service_v1 import OllamaServiceV1
-from src.api.v1.schemas.generate import GenerateRequest, GenerateResponse
+from src.api.v1.schemas import GenerateRequest, GenerateResponse
 
 router = APIRouter(
     prefix="/api/v1",
