--- src/api/v1/schemas/__init__.py
@@ -1,7 +1,0 @@
-"""
-Schemas for v1 API endpoints.
-"""
-
-from .generate import GenerateRequest, GenerateResponse
-
-__all__ = ["GenerateRequest", "GenerateResponse"]
