--- src/api/v2/services/__init__.py
@@ -1,7 +1,0 @@
-"""
-Services for v2 API endpoints.
-"""
-
-from ..ollama_service_v2 import OllamaServiceV2
-
-__all__ = ["OllamaServiceV2"]
